+++ src/server/game/World/World.cpp
@@ -2876,7 +2876,7 @@ void World::LoadDBAllowedSecurityLevel()
     PreparedQueryResult result = LoginDatabase.Query(stmt);
     
     if (result)
-        SetPlayerSecurityLimit(AccountTypes(result->Fetch()->GetUInt16()));
+        SetPlayerSecurityLimit(AccountTypes(result->Fetch()->GetUInt8()));
 }
 
 void World::SetPlayerSecurityLimit(AccountTypes _sec)
